backend_fastapi/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy[asyncio]==2.0.43
asyncpg==0.30.0
pydantic-settings==2.10.1
python-dotenv==1.1.1
